Restructure models
