+++ app/src/main/java/com/hieuwu/groceriesstore/domain/usecases/AddMealToPlanUseCase.kt
@@ -1,10 +1,11 @@
 package com.hieuwu.groceriesstore.domain.usecases
 
-import com.hieuwu.groceriesstore.presentation.mealplanning.overview.MealType
+import com.hieuwu.groceriesstore.presentation.mealplanning.overview.state.MealType
 
 interface AddMealToPlanUseCase :
     SuspendUseCase<AddMealToPlanUseCase.Input, AddMealToPlanUseCase.Output> {
     class Input(val name: String, val weekDay: String, val ingredients: List<String>,
-        val mealType: MealType)
+        val mealType: MealType
+    )
     object Output
 }
