+++ app/src/main/java/com/hieuwu/groceriesstore/domain/usecases/RetrieveMealByTypeUseCase.kt
@@ -1,8 +1,8 @@
 package com.hieuwu.groceriesstore.domain.usecases
 
 import com.hieuwu.groceriesstore.domain.models.MealModel
-import com.hieuwu.groceriesstore.presentation.mealplanning.overview.MealType
-import com.hieuwu.groceriesstore.presentation.mealplanning.overview.WeekDayValue
+import com.hieuwu.groceriesstore.presentation.mealplanning.overview.state.MealType
+import com.hieuwu.groceriesstore.presentation.mealplanning.overview.state.WeekDayValue
 
 interface RetrieveMealByTypeUseCase :
     SuspendUseCase<RetrieveMealByTypeUseCase.Input, RetrieveMealByTypeUseCase.Output> {
