+++ app/src/main/java/com/hieuwu/groceriesstore/presentation/mealplanning/overview/state/Meal.kt
@@ -1,4 +1,4 @@
-package com.hieuwu.groceriesstore.presentation.mealplanning.overview
+package com.hieuwu.groceriesstore.presentation.mealplanning.overview.state
 
 data class Meal(
     val id: String,
