+++ app/src/main/java/com/hieuwu/groceriesstore/presentation/mealplanning/overview/state/MealType.kt
@@ -1,4 +1,4 @@
-package com.hieuwu.groceriesstore.presentation.mealplanning.overview
+package com.hieuwu.groceriesstore.presentation.mealplanning.overview.state
 
 enum class MealType(val value: String) {
     LUNCH("lunch"), BREAKFAST("breakfast"), DINNER("dinner")
