+++ app/src/main/java/com/hieuwu/groceriesstore/presentation/mealplanning/overview/state/WeekDayValue.kt
@@ -1,4 +1,4 @@
-package com.hieuwu.groceriesstore.presentation.mealplanning.overview
+package com.hieuwu.groceriesstore.presentation.mealplanning.overview.state
 
 enum class WeekDayValue(val dayValue: String) {
     Mon("Mon"),
